Remove all code relating to the legacy tenant app.
