+++ frontend/lib/norent/start-account-or-login/routes.ts
@@ -11,7 +11,6 @@ export function createStartAccountOrLoginRouteInfo(prefix: string) {
   return {
     phoneNumber: `${prefix}/phone/ask`,
     phoneNumberTermsModal: `${prefix}/phone/ask/terms-modal`,
-    migrateLegacyTenantsUser: `${prefix}/migrate-legacy-tenants-user`,
     verifyPhoneNumber: `${prefix}/phone/verify`,
     setPassword: `${prefix}/password/set`,
     verifyPassword: `${prefix}/password/verify`,
