--- frontend/lib/queries/autogen/PrepareLegacyTenantsAccountForMigrationMutation.graphql
@@ -1,7 +1,0 @@
-# This file was auto-generated by querybuilder, please do not edit it.
-mutation PrepareLegacyTenantsAccountForMigrationMutation($input: PrepareLegacyTenantsAccountForMigrationInput!) {
-  output: prepareLegacyTenantsAccountForMigration(input: $input) {
-    errors { ...ExtendedFormFieldErrors },
-    session { ...AllSessionInfo }
-  }
-}
