--- legacy_tenants/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class LegacyTenantsConfig(AppConfig):
-    name = 'legacy_tenants'
