+++ mypy.ini
@@ -28,8 +28,5 @@ ignore_errors = True
 [mypy-issues.tests.*]
 ignore_errors = True
 
-[mypy-legacy_tenants.tests.*]
-ignore_errors = True
-
 [mypy-nycha.tests.*]
 ignore_errors = True
