+++ project/justfix_environment.py
@@ -60,15 +60,6 @@ class JustfixEnvironment(typed_environ.BaseEnvironment):
     # production-like setups in a jiffy.
     USE_DEVELOPMENT_DEFAULTS: bool = False
 
-    # This is the URL to the MongoDB instance of the legacy
-    # tenants app, e.g. "mongodb://localhost:27017/somedb". If undefined,
-    # connectivity to the legacy tenants app will be disabled.
-    LEGACY_MONGODB_URL: str = ''
-
-    # This is the URL for the origin of the legacy tenants app,
-    # e.g. "https://beta.justfix.nyc".
-    LEGACY_ORIGIN: str = 'https://beta.justfix.nyc'
-
     # This is an optional HTTP request header field name and
     # value indicating that the request is actually secure.
     # For example, Heroku deployments should set this to
