+++ users/admin.py
@@ -6,8 +6,6 @@ from .forms import JustfixUserCreationForm, JustfixUserChangeForm
 from .models import JustfixUser
 import rapidpro.models
 from onboarding.admin import OnboardingInline
-from legacy_tenants.admin import LegacyUserInline
-from legacy_tenants.models import LegacyUserInfo
 from .admin_user_proxy import user_signup_intent, sms_conversations_field
 from texting.models import get_lookup_description_for_phone_number
 from loc.admin import LOCUser, LandlordDetailsInline
@@ -91,7 +89,6 @@ class JustfixUserAdmin(airtable.sync.SyncUserOnSaveMixin, UserAdmin):
         }),
     )
     inlines = (
-        LegacyUserInline,
         OnboardingInline,
         # We'll consider this part of the core user info b/c all our
         # flows ask for it, and it can be considered part of the user's
@@ -117,14 +114,6 @@ class JustfixUserAdmin(airtable.sync.SyncUserOnSaveMixin, UserAdmin):
         if obj is not None and not request.user.is_superuser:
             return self.non_superuser_fieldsets
         return super().get_fieldsets(request, obj)
-
-    def get_formsets_with_inlines(self, request, obj=None):
-        for inline in self.get_inline_instances(request, obj):
-            # Don't show the legacy user inline if they're not a legacy user.
-            if (isinstance(inline, LegacyUserInline) and
-                    not LegacyUserInfo.is_legacy_user(obj)):
-                continue
-            yield inline.get_formset(request, obj), inline
 
     hp_action_info = make_link_to_other_user_view(HPUser, "HP action information")
